refactor(widget): Clean up waveform widget
